Cleanup spacing.  Move worker out of directory since it’s not a module
