+++ app/controllers/job_test_controller.rb
@@ -1,18 +1,16 @@
-require_relative '../jobs/killabeez.rb'
-require_relative '../jobs/docker.rb'
-
 
 class JobTestController < ApplicationController
-    def swarm
-        KillaBeez.create(:cmd => 'sleep 5')
+  def swarm
+    KillaBeez.create(:cmd => 'sleep 5')
 
-        render text: "KillaBeez on the loose!"
-    end
-    def docker
-      DockerWorker.create(
-        :url => 'https://github.com/performanceci/simple.git',
-        :repo => 'simple'
-      )
-      render text: "Docker is better"
-    end
+    render text: "KillaBeez on the loose!"
+  end
+
+  def docker
+    DockerWorker.create(
+      :url => 'https://github.com/performanceci/simple.git',
+      :repo => 'simple'
+    )
+    render text: "Docker is better"
+  end
 end
